\name{ALimage}
\Rdversion{1.1}
\alias{ALimage}
\title{
 Show Functional Antennal Lobe 
}
\description{
  Show the functional antennal lobe responded to a single odor.
}
\usage{
ALimage(response.data, tag.ALimage = default.val("tag.ALimage"), main = default.val("main"), col.extrem = default.val("col.extrem"), ind.panels = default.val("ind.panels"), OGN, AL256)
}
\arguments{
  \item{response.data}{
a data frame; normalized response data across receptors, which is generated by \code{\link{findRespNorm}}.
}
  \item{tag.ALimage}{
character string; Default is "Glomerulus", the options include "Ors" and "ORN".
}
  \item{main}{
title of figure; Default is NULL.
}
  \item{col.extrem}{
character vector;  The default is c("blue", "red"), i.e. the color scale ranges from blue to red 
}
  \item{ind.panels}{
logical; if \code{TRUE}, the indicated panels are shown on the right side.
}
  \item{OGN}{
a data frame; maps receptor to sensillum, receptor neuron and glomerulus.
}
  \item{AL256}{
a data frame; which contains pixel values to code the graphic antennal lobe.
}
}
\details{
normalized response data across receptors to a given odor can obtained using \code{\link{findRespNorm}}. 
      The function then maps the receptor responses onto the respective glomeruli, resulting in a colored map of the
      antennal lobe. Default colors range from red (most inhibitory response) over white (around baseline) to blue (most excitatory response), see \code{\link{DoORColmap}}.
      
      The three-dimensional antennal lobe is displayed as four two-dimensional slices at different z-positions. 
      Light grey glomeruli shine through from other slices (background, BG).  
      Unmapped (UM) glomeruli are painted in dark grey - in these cases the receptor-glomerulus mapping is currently unknown. 
      Grey indicates NA, i.e. that no measurements are available for the respective glomerulus (receptor).
}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>
}
\seealso{
\code{\link{findRespNorm}},\code{\link{CreateDatabase}}
}
\examples{
library(DoOR.data)
library(DoOR.function)
loadRD()
data(response.range)
data(OGN)
data(AL256)
cas<-c("628-63-7")
RP.PAcE<-findRespNorm(cas,responseMatrix= response.matrix)
ALimage(RP.PAcE, main = "PAcE", OGN = OGN, AL256 = AL256)
}
\keyword{ hplot }
