% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calModel.R
\name{calModel}
\alias{calModel}
\title{' select the best model function}
\usage{
calModel(x, y, select.MD = default.val("select.MD"))
}
\arguments{
\item{x,y}{data vectors from study x and y (can contain NA)}

\item{select.MD}{logical, if TRUE, only the best model function (in terms of
MD) will be returned.}
}
\description{
\code{calModel} is used to return the best model function that represent the
relationship between responses from study x and y.
}
\details{
\code{calModel} chooses the best model function from following: linear,
exponential function, sigmoid, asymptotic model with x intercept, asympototic
model with y intercept and their inverse versions. (If your are interested in
these functions please check the sources at
https://github.com/Dahaniel/DoOR.functions)
}
\examples{
library(DoOR.data)
data(Or35a)
x <- DoORnorm(Or35a[,6])
y <- DoORnorm(Or35a[,8])
calM_xy <- calModel(x, y, select.MD = default.val("select.MD"))
}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>
}

