% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combData.R
\name{combData}
\alias{combData}
\title{Combine Two Data}
\usage{
combData(data1, data2, by.data2, assigned.name, ident = default.val("ident"))
}
\arguments{
\item{data1}{data frame; the object to be added.}

\item{data2}{data frame; it will be merged to "data1".}

\item{by.data2}{character string; a column name in "data2", specify which
column will be combined into data1.}

\item{assigned.name}{charater string; assigning a name to combined column.}
}
\description{
Combine two data into one byInChIKey
}
\details{
Data frames are merged by InChIKey.
Data2 will be merged into data1 specified by the column name \code{by.data2}
into one. The name of added column could be assigned with
\code{assigned.name}. If \code{data2} contains some odors that the
\code{data1} do not have, then the new odors will be added into output.
}
\examples{
library(DoOR.data)
library(DoOR.function)
newdata <- data.frame(InChIKey=odor$InChIKey[1:4],
	new.data=c(0.4, 0.5, 0.6, 0.7))
data(Or22a)
new.Or22a <- combData(data1 = Or22a, data2 = newdata, by.data2 = "new.data")
}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>
}
\keyword{data}

