% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_studies.R
\name{count_studies}
\alias{countStudies}
\alias{count_studies}
\title{count_studies}
\usage{
count_studies(ors = door_default_values("ORs"),
  odor_data = door_default_values("odor"),
  char.columns = door_default_values("num.charColumns"),
  ident = door_default_values("ident"))
}
\arguments{
\item{ors}{data.frame containing all receptors exidting in DoOR.}

\item{odor_data}{data.frame containing information about the odorants in
DoOR.}

\item{char.columns}{number of character columns in each receptor data.frame.}

\item{ident}{odorant identifier to be used as rownames.}
}
\value{
matrix
}
\description{
returns a matrix indiating how many studies have recorded individual
receptor-odorant combinations
}
\examples{
library(DoOR.data)
load_door_data()
count <- countStudies()
image(count)

}

