% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/countStudies.R
\name{countStudies}
\alias{countStudies}
\title{countStudies}
\usage{
countStudies(ors = default.val("ORs"), odor.data = default.val("odor.data"),
  char.columns = default.val("num.charColumns"),
  ident = default.val("ident"))
}
\arguments{
\item{ors}{data.frame containing all receptors exidting in DoOR.}

\item{odor.data}{data.frame containing information about the odorants in DoOR.}

\item{char.columns}{number of character columns in each receptor data.frame.}

\item{ident}{odorant identifier to be used as rownames.}
}
\value{
matrix
}
\description{
returns a matrix indiating how many studies have recorded individual receptor-odorant combinations
}
\examples{
count <- countStudies()
image(count)
}

