% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_door_database.R
\name{create_door_database}
\alias{CreateDatabase}
\alias{create_door_database}
\title{Compose a Response Matrix of All Odor Receptors}
\usage{
create_door_database(tag = door_default_values("tag"),
  select.MDValue = door_default_values("select.MDValue"),
  overlapValues = door_default_values("overlapValues"), ...)
}
\arguments{
\item{tag}{character string, format for rownames, possibilities: "InChIKey",
CAS", "CID", "Name"}

\item{select.MDValue}{a numeric, threshold on the MD, this is used to reject
studies that do not align sufficiently well to the response model}

\item{overlapValues}{numeric, a criterion using to refuse a data set that
has not enough overlap value.}

\item{...}{pass more parameters to \code{\link{model_response}}}
}
\description{
computes the complete response model for all receptors in the database
(calls \code{\link{model_response}} for all receptors)
}
\examples{
\dontrun{
library(DoOR.data)
load_door_data()
mydatabase <- create_door_database()
}
}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>

Daniel Münch <\email{daniel.muench@uni-konstanz.de}>
}
\seealso{
\code{\link{model_response}}
}
\keyword{data}

