% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.functions.version.R
\name{DoOR.functions.version}
\alias{DoOR.functions.version}
\title{DoOR.functions package version}
\usage{
DoOR.functions.version()
}
\description{
Display the DoOR.functions package version when loading.
}
\keyword{DoOR}

