% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/door_default_values.R
\name{door_default_values}
\alias{default.val}
\alias{door_default_values}
\title{default values for DoOR functions}
\usage{
door_default_values(DoOR_default)
}
\arguments{
\item{DoOR_default}{a character string, indicating which argument is to be
returned for DoOR functions.}
}
\description{
\code{door_default_values} is used to return default values for DoOR functions.
}
\details{
There are six categories for default value. real number, integer, logical,
NULL, character string and character vector.
}
\examples{
# extract DoOR default values
door_default_values(DoOR_default = "select.MD")

}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>
}
\keyword{data}

