% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoORmelt.R
\name{DoORmelt}
\alias{DoORmelt}
\title{DoORmelt}
\usage{
DoORmelt(data, datasets, ident, na.rm = F)
}
\arguments{
\item{data}{data frame containing either subsets of response.matrix or identifier and data columns (e.g. Or22a)}

\item{datasets}{column names of data sets to use, important when using response data (e.g. Or22a); if empty all columns will be used}

\item{ident}{chemical identifier, if empty rownames of data will be used}

\item{na.rm}{should NAs be removed}
}
\value{
"long" data frame for plotting with ggplot2
}
\description{
prepare odor response data for plotting via ggplot2
}
\examples{
head(DoORmelt(Or22a, datasets = c("Hallem.2004.WT", "Pelz.2006.AntEC50")))
head(DoORmelt(response.matrix[1:100,], na.rm = T))
}
\author{
Daniel Münch <\email{daniel.muench@uni-konstanz.de}>
}

