% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoORnorm.R
\name{DoORnorm}
\alias{DoORnorm}
\title{rescale the data values from 0 to 1}
\usage{
DoORnorm(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
\code{DoORnorm} is used to normalize the data in values from 0 to 1.
}
\examples{
x <- rnorm(10)
DoORnorm(x)
}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>
}
\keyword{math}

