% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoORplot_acrossReceptors.R
\name{DoORplot_acrossReceptors}
\alias{DoORplot_acrossReceptors}
\title{DoORplot_acrossReceptors}
\usage{
DoORplot_acrossReceptors(odorants,
  responseMatrix = default.val("response.matrix"),
  zero = default.val("zero"), tag = "Name", limits)
}
\arguments{
\item{odorants}{character vecto; one or several InChIKeys}

\item{responseMatrix}{DoOR response matrix; a DoOR response matrix as data source}

\item{zero}{character; an InChIKey of the odorant that should be set to 0}

\item{tag}{character; the chemical identifier to plot as odorant name (one of colnames(odor))}

\item{limits}{numeric of length 2; if provided the ylim will range accordingly}
}
\value{
a ggplot object
}
\description{
barplot of DoOR responses of a set of odorants across all responding units in DoOR
}
\examples{
DoORplot_acrossReceptors(getKey("123-92-2"), tag = "CAS")
DoORplot_acrossReceptors(odor$InChIKey[4:10])
}

