% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoORplot_ALmap.R
\name{DoORplot_ALmap}
\alias{DoORplot_ALmap}
\title{DoORplot_ALmap}
\usage{
DoORplot_ALmap(InChIKey, responseMatrix = default.val("response.matrix"),
  zero = default.val("zero"), tag = default.val("tag.DoORplot_ALmap"),
  main = "Name", scalebar = default.val("scalebar"),
  DoOR.mappings = default.val("DoOR.mappings"),
  AL.map = default.val("AL.map"), colors = c("#0570b0", "#74a9cf",
  "#ffffff", "#fdcc8a", "#fc8d59", "#d7301f"), legend = T, limits)
}
\arguments{
\item{InChIKey}{InChIKey specifying the odorant to plot}

\item{responseMatrix}{the input data (e.g. response.matrix or response.matrix_non.normalized)}

\item{zero}{the odorant to set to zero (defaults to "SFR")}

\item{tag}{the labels to plot on top of the glomeruli (one of the following \code{DoOR.mappings} columns: "receptor", "sensillum", "ORN", "glomerulus" or "co.receptor")}

\item{main}{the title, one column of \code{odor}, defaults to "Name"}

\item{scalebar}{whether or not to add a scalebar}

\item{DoOR.mappings}{the darta frame containing the mapping information}

\item{colors}{a vector containing 6 color values (2 for values below 0, 1 0 value and 3 steps between 0 and 1)}

\item{limits}{the limits for the color scale, if empty the range of the response matrix is taken (after setting ``zero`` to 0)}
}
\value{
a ggplot2 object
}
\description{
Plot an antennal lobe map with color coded odorant responses.
}
\details{
Normalized, color coded odor responses across receptors are mapped onto a map of the \emph{Drosophila} antennal lobe. The antennal lobe map was a kind gift from Veit Grabe.
}
\examples{
DoORplot_ALmap("MLFHJEHSLIIPHL-UHFFFAOYSA-N", scalebar = F)
DoORplot_ALmap("MLFHJEHSLIIPHL-UHFFFAOYSA-N", tag = "Ors", color = c("magenta", "pink", "white", "yellow", "orange", "red"))

DoORplot_ALmap(getKey("123-92-2"), scalebar = F) +
theme(legend.position  = "bottom",
      panel.background = element_rect(fill = "grey90", color = NA)) +
ggtitle("responses elicited by isopentyl acetate")

\dontrun{
p <- DoORplot_ALmap(getKey("123-92-2"))
ggsave("AL.response.pdf", p, width = 6, height = 2, scale = 2)
}
}
\author{
Daniel Münch \email{daniel.muench@uni-konstanz.de}
}
\references{
Grabe, V., Strutz, A., Baschwitz, A., Hansson, B.S., Sachse, S., 2014. A digital in vivo 3D atlas of the antennal lobe of Drosophila melanogaster. J. Comp. Neurol. n/a–n/a. doi:10.1002/cne.23697
}
\seealso{
\link{findRespNorm}, \pkg{\link{ggplot2}, \link{grid}}
}

