% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoORplot_responseProfile.R
\name{DoORplot_responseProfile}
\alias{DoORplot_responseProfile}
\title{DoORplot_responseProfile}
\usage{
DoORplot_responseProfile(data, receptor, tag = default.val("tag"),
  colored = T, colors = default.val("colors"), limits,
  scalebar = default.val("scalebar"))
}
\arguments{
\item{data}{a DoOR response.matrix}

\item{receptor}{character string, receptor name, any of colnames(response.matrix)}

\item{tag}{chemical identifier for annotation}

\item{colored}{color code the bars according to the response value?}

\item{colors}{a vector of 5 colors (2 for values < 0, 1 value for 0 and 3 values > 0)}

\item{limits}{the limits for the colorscale and the x axis, global range of data will be used if empty}

\item{scalebar}{add or suppress scalebars}
}
\value{
ggplot2 plot
}
\description{
create a barplot of a DoOR response profile
}
\examples{
DoORplot_responseProfile(response.matrix, "Or22a")
}

