% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoORplot_tuningCurve.R
\name{DoORplot_tuningCurve}
\alias{DoORplot_tuningCurve}
\title{Receptor tuning breadth}
\usage{
DoORplot_tuningCurve(x, ...)
}
\arguments{
\item{x}{a numeric vector}

\item{\dots}{further graphical parameters, see \code{\link{par}}}
}
\description{
Display the distribution of a receptor's responses (pyramid plot).
}
\details{
This function is used to visualise the response spectrum of a receptor.
Generalists have a distribution with broad tails, while specialist receptors
have a distinct peak that decreases sharply towards the tails.
}
\examples{
library(DoOR.data)
data(Or22a)
x=Or22a[,"Hallem.2006.EN"]
x<-na.omit(x)
DoORplot_tuningCurve(x,las=2,main="Tuning Curve:Hallem.2006.EN")
}
\references{
Hallem E. A. & Carlson J. R., 2006, \emph{Coding of odors by a
receptor repertoire}, Cell, 125:143-160 \cr
\url{http://www.cell.com/content/article/abstract?uid=PIIS0092867406003631}
}
\seealso{
\code{\link{barplot}}
}
\keyword{iplot}

