% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot_across_ru.R
\name{dplot_across_ru}
\alias{dplot_acrossReceptors}
\alias{dplot_across_ru}
\title{dplot_across_ru}
\usage{
dplot_across_ru(odorant,
  response_matrix = door_default_values("door_response_matrix"),
  odor_data = door_default_values("odor"),
  zero = door_default_values("zero"), tag = "Name", limits,
  base_size = 12)
}
\arguments{
\item{odorant}{character vecto, one or several InChIKeys}

\item{response_matrix}{DoOR response matrix, a DoOR response matrix as data source}

\item{odor_data}{data frame, contains the odorant information.}

\item{zero}{character, an InChIKey of the odorant that should be set to 0}

\item{tag}{character, the chemical identifier to plot as odorant name (one of colnames(odor))}

\item{limits}{numeric of length 2, if provided the ylim will range accordingly}

\item{base_size}{numeric, the base font size for the ggplot2 plot}
}
\value{
a ggplot object
}
\description{
barplot of DoOR responses of a set of odorant across all responding units in DoOR
}
\examples{
# load data
library(DoOR.data)
library(DoOR.functions)
data(odor)

# plot activation pattern of one or several odorants
dplot_across_ru(trans_id("123-92-2"), tag = "CAS")
dplot_across_ru(odor$InChIKey[4:10])

}
\author{
Daniel Münch <\email{daniel.muench@uni-konstanz.de}>
}

