% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot_response_matrix.R
\name{dplot_response_matrix}
\alias{dplot_response_matrix}
\alias{dplot_responseMatrix}
\title{dplot_response_matrix}
\usage{
dplot_response_matrix(data, odor_data = door_default_values("odor"),
  tag = door_default_values("tag"), colors = door_default_values("colors"),
  flip = FALSE, fix = TRUE, bw = FALSE, point = FALSE, limits,
  base_size = 12)
}
\arguments{
\item{data}{a subset of e.g. door_response_matrix}

\item{odor_data}{data frame, contains the odorant information.}

\item{tag}{the chemical identfier to plot (one of colnames(odor))}

\item{colors}{the colors to use if negative values are supplied (range of 5
colors, 2 for negative values, 1 for 0 and 3 for positive values)}

\item{flip}{logical, if TRUE the x and y axes will be flipped}

\item{fix}{logical, whether to fix the ratio of the tiles when plotting as a heatmap}

\item{bw}{logical, whether to plot b&w or colored}

\item{point}{logical, if \code{TRUE} a point matrix instead of a heatmap will
be returned (the default if you supply only positive values)}

\item{limits}{the limits of the scale, will be calculated if not set}

\item{base_size}{numeric, the base font size for the ggplot2 plot}
}
\value{
a dotplot if limits[1] >= 0  or a heatmap if limits[1] < 0
}
\description{
plot DoOR responses as a point matrix
}
\examples{
# load data
library(DoOR.data)
data(door_response_matrix)

# reset the spontaneous firing rate to 0
tmp <- reset_sfr(door_response_matrix, "SFR")

# plot heatmap / coloured tiles
dplot_response_matrix(tmp[10:50,], tag = "Name", limits = range(tmp, na.rm = TRUE))

# plot dotplot
dplot_response_matrix(door_response_matrix[10:50,], tag = "Name",
                        limits = range(door_response_matrix, na.rm = TRUE))

}
\author{
Daniel Münch <\email{daniel.muench@uni-konstanz.de}>
}
