% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot_tuningCurve.R
\name{dplot_tuningCurve}
\alias{dplot_tuningCurve}
\title{dplot_tuningCurve}
\usage{
dplot_tuningCurve(receptor, odorant, response.vector,
  response_matrix = door_default_values("response.matrix"),
  odor_data = door_default_values("odor"),
  zero = door_default_values("zero"),
  fill.receptor = door_default_values("color.receptor"),
  fill.odorant = door_default_values("color.odorant"), odor.main = "Name",
  limits, base_size = 12)
}
\arguments{
\item{receptor}{character, a receptor name (one of ORS$OR)}

\item{odorant}{character, an odorant name (InChIKey)}

\item{response.vector}{numerical vector, a vector with responses, if empty
this is taken from response.matrix}

\item{response_matrix}{DoOR response matrix, response vector will be taken
from here, not needed if response.vector is given}

\item{odor_data}{data frame, contains the odorant information.}

\item{zero}{InChIKey, will be set to zero, default is SFR, ignored when data
is provided via response.vector, set to "" if you don't want to subtract
anything}

\item{fill.receptor}{color code, bar color for receptor tuning curve}

\item{fill.odorant}{color code, bar color for odorant tuning curve}

\item{odor.main}{the odor identifier to plot, one of colnamed(odor)}

\item{limits}{the numerical vector of length 2, y limits for the tuning curve}

\item{base_size}{numeric, the base font size for the ggplot2 plot}
}
\value{
a ggplot object
}
\description{
plot a receptor or odorant tuning curve
}
\examples{
library(DoOR.data)
data(response.matrix)
data(odor)

dplot_tuningCurve(odorant = odor$InChIKey[2])
dplot_tuningCurve(receptor = "Or22a")

range <- range(reset_sfr(response.matrix, "SFR"), na.rm = TRUE)
dplot_tuningCurve(receptor = "Or10a", limits = range, fill.receptor = "magenta")

dplot_tuningCurve(receptor = "OrX", response.vector = c(1:100))

dplot_tuningCurve(odorant = "odor X", response.vector = rnorm(200))

}
\author{
Daniel Münch <\email{daniel.muench@uni-konstanz.de}>
}

