% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dsdx_asymp_inverse.R
\name{dsdx_asymp_inverse}
\alias{dsdx_asymp_inverse}
\title{Leibniz's notation for computing the curvic length on inverse Asymptotic
Model}
\usage{
dsdx_asymp_inverse(input, parms)
}
\arguments{
\item{input}{numeric vector of data values}

\item{parms}{numeric vector; parameters with given names: "Asym", "lrc",
"R0"}
}
\description{
Leibniz's notation for computing the curvic length on inverse Asymptotic
Model
}
\details{
expression : sqrt(1+(-(1/(R0 - Asym)/(input - Asym)/(R0 -
Asym)/exp(lrc)))^2)
}
\examples{
x <-  seq(0.1,1,length=20)
parms <- c(Asym = 2, R0 = 0.2, lrc = 0.3)
integrate(function(x) { dsdx_asymp_inverse(input = x, parms = parms ) }, lower = 0, upper = 0.5)
}
\references{
Jose Pinheiro and Douglas Bates
}
\seealso{
\code{\link{SSasymp}}
}
\keyword{math}

