% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dsdx_exp_inverse.R
\name{dsdx_exp_inverse}
\alias{dsdx_exp_inverse}
\title{Leibniz's notation for computing the curvic length on inverse exponential
model}
\usage{
dsdx_exp_inverse(input, parms)
}
\arguments{
\item{input}{numeric vector of data values}

\item{parms}{numeric vector; parameters with given names: "a", "b", "c"}
}
\description{
Leibniz's notation for computing the curvic length on inverse exponential
model
}
\details{
expression : sqrt(1 + (1/b/((input - a)/b)/c)^2)
}
\examples{
x <-  seq(0.1,1,length=20)
parms <- c(a = 0, b = 0.2, c = 0.3)
integrate(function(x) { dsdx_exp_inverse(input = x, parms = parms ) }, lower = 0.1, upper = 0.5)
}
\references{
Jose Pinheiro and Douglas Bates
}
\seealso{
\code{\link{SSexpo}}
}
\keyword{math}

