% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dsdx_linear_inverse.R
\name{dsdx_linear_inverse}
\alias{dsdx_linear_inverse}
\title{Leibniz's notation for computing the curvic length on inverse linear model}
\usage{
dsdx_linear_inverse(input, parms)
}
\arguments{
\item{input}{numeric vector of data values}

\item{parms}{numeric vector; parameters with given names: "Intercept",
"Slope"}
}
\description{
Leibniz's notation for computing the curvic length on inverse linear model
}
\details{
expression : sqrt(1+(1/Slope)^2)*input^0
}
\examples{
x <-  seq(0.1,1,length=20)
parms <- c(Intercept = 0.2, Slope = 0.3)
integrate(function(x) { dsdx_linear_inverse(input = x, parms = parms ) }, lower = 0.1, upper = 0.5)
}
\keyword{math}

