% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dsdx_sigmoid_inverse.R
\name{dsdx_sigmoid_inverse}
\alias{dsdx_sigmoid_inverse}
\title{Leibniz's notation for computing the curvic length on inverse Logistic Model}
\usage{
dsdx_sigmoid_inverse(input, parms)
}
\arguments{
\item{input}{numeric vector of data values}

\item{parms}{numeric vector; parameters with given names: "Asym", "xmid",
"scal"}
}
\description{
Leibniz's notation for computing the curvic length on inverse Logistic Model
}
\details{
expression : sqrt(1+(scal * (Asym/input^2/(Asym/input - 1)))^2)
}
\examples{
x <-  seq(0.1,1,length=20)
parms <- c(Asym = 2, xmid = 0.2, scal = 0.3)
integrate(function(x) { dsdx_sigmoid_inverse(input = x, parms = parms ) }, lower = 0.1, upper = 0.5)
}
\references{
Jose Pinheiro and Douglas Bates
}
\seealso{
\code{\link{SSlogis}}
}
\keyword{math}

