% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_missing_value.R
\name{estimate_missing_value}
\alias{DoOREst}
\alias{estimate_missing_value}
\title{Estimate the missing entries in a response data}
\usage{
estimate_missing_value(data, nodor, method = "PC")
}
\arguments{
\item{data}{a data frame or matrix, contaning the consensus response values}

\item{nodor}{a numeric value, specifying the number of the selected odors}

\item{method}{character string, specifying the method ("PC" (Pearson's
coefficient) and "Knn" (k nearest neighbors)) for estimation, the default is
"PC".}
}
\description{
Estimate the missing entries in a response data
}
\details{
A wrapper programe for using Pearson Correlation or k-nearest neighbors to
estimate the missing entries in a response matrix.
}
\examples{
\dontrun{
library(DoOR.data)
data(response.matrix)
subset <- response.matrix[1:100, 11:30]
est.data <- estimate_missing_value(data = subset, nodor = 6)
}
}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>
}
\references{
Kim, H., Golub, G. H. & Park, H., Missing value estimation for
DNA microarray gene expression data: local least squares imputation., 2005,
Bioinformatics, 21, 187-198
}
\keyword{math}

