% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_door_data.R
\name{export_door_data}
\alias{export_door_data}
\title{export data}
\usage{
export_door_data(directory, sep = ";", ...)
}
\arguments{
\item{directory}{character string, output dir}

\item{sep}{separator used in write.csv}

\item{...}{more parameters passed to write.csv}
}
\description{
export odor response data and supported data
}
\details{
Please load ORs from data package DoOR.data by typing (\code{data(ORs)}) 
before use.
}
\examples{
\dontrun{
# load data
library(DoOR.data)
library(DoOR.functions)
load_door_data()

# export odorant response data only
export_door_data(".") 	
}
}
\author{
Daniel Münch <\email{daniel.muench@uni-giessen.de}>
}
\keyword{data}
