% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filterData.R
\name{filterData}
\alias{filterData}
\title{library(DoOR.data)}
\usage{
filterData(data, overlapValues = default.val("overlapValues"),
  charColumns = default.val("num.charColumns"))
}
\description{
library(DoOR.data)
}

