% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findResp.R
\name{findResp}
\alias{findResp}
\title{Find receptor responses}
\usage{
findResp(odorant, responseRange = default.val("response.range"),
  Or.list = loadRDList())
}
\arguments{
\item{odorant}{a single odor provided as InChIKey}

\item{responseRange}{data frame; response ranges of studies}

\item{Or.list}{a list contains reponse data of all available receptors. It
can be loaded using \code{\link{loadRDList}}.}
}
\description{
given a chemical, get original receptor responses from all studies in the
database.
}
\details{
output is a data frame containing response values of given odor across
receptors from all available studies.
}
\examples{
library(DoOR.data)
loadRD()
responses <- findResp(odor = 'MLFHJEHSLIIPHL-UHFFFAOYSA-N')
}
\keyword{data}

