% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_normalized_responses.R
\name{get_normalized_responses}
\alias{getNormalizedResponses}
\alias{get_normalized_responses}
\title{Find normalised receptor responses}
\usage{
get_normalized_responses(odors, zero = door_default_values("zero"),
  response_matrix = door_default_values("door_response_matrix"), round = 3,
  na.rm = FALSE)
}
\arguments{
\item{odors}{character vector, one or more odors provided as InChIKey.}

\item{zero}{InChIKey of background that should be set to zero. The default is
"SFR", i.e. the spontaneous firing rate.}

\item{response_matrix}{a data frame, as e.g. "door_response_matrix" that is loaded
by \code{\link{model_response}}. It is also possible to create this frame manually
using \code{\link{model_response}}.}

\item{round}{numeric, round to this amount of digits, set to NA if you do not want to round}

\item{na.rm}{logical, remove NAs?}
}
\description{
given a chemical, get normalised receptor responses from all studies in the
database.
}
\examples{
# load data
library(DoOR.data)
data(door_response_matrix)

# define a list of odorants
odors <- c("MLFHJEHSLIIPHL-UHFFFAOYSA-N",
           "OBNCKNCVKJNDBV-UHFFFAOYSA-N",
           "IKHGUXGNUITLKF-UHFFFAOYSA-N")

# get the normalized responses for these odorants
result <- get_normalized_responses(odors, response_matrix = door_response_matrix)

}
\author{
Daniel Münch \email{daniel.muench@uni-konstanz.de}
}
\seealso{
\code{\link{model_response}},\code{\link{create_door_database}}
}
\keyword{data}

