% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_responses.R
\name{get_responses}
\alias{getResponses}
\alias{get_responses}
\title{Find receptor responses}
\usage{
get_responses(odorant,
  responseRange = door_default_values("door_response_range"),
  Or.list = load2list())
}
\arguments{
\item{odorant}{a single odor provided as InChIKey}

\item{responseRange}{data frame, response ranges of studies}

\item{Or.list}{a list contains reponse data of all available receptors. It
can be loaded using \code{\link{load2list}}.}
}
\description{
given a chemical, get original receptor responses from all studies in the
database.
}
\details{
output is a data frame containing response values of given odor across
receptors from all available studies.
}
\examples{

library(DoOR.data)
load_door_data()
responses <- get_responses(odor = 'MLFHJEHSLIIPHL-UHFFFAOYSA-N')

}
\author{
Daniel Münch \email{daniel.muench@uni-konstanz.de}
}
\keyword{data}

