% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataset.R
\name{getDataset}
\alias{getDataset}
\title{getDataset}
\usage{
getDataset(study, na.rm = FALSE)
}
\arguments{
\item{study}{character, the name of the study you want to aggregate the dta 
from}

\item{na.rm}{logical, whether or not to exclude odorants that were not 
measured in the study}
}
\value{
returns a data frame containing all the odorant responses measured in
  \code{study}
}
\description{
aggregates original data from a given study
}
\examples{
loadData()
getDataset("Hallem.2004.EN", na.rm = TRUE)
}
\author{
Daniel Münch <\email{daniel.muench@uni-konstanz.de}>
}

