% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getKey.R
\name{getKey}
\alias{getKey}
\title{getKey}
\usage{
getKey(x, type = "CAS", odor.data = default.val("odor.data"))
}
\arguments{
\item{x}{character, an odor identifier}

\item{type}{the type od identifier (one of the column names of ``odor``)}

\item{odor.data}{the data frame containing the odor information (defaults to ``odor``).}
}
\value{
character, an InChIKey
}
\description{
Translate an odor identifier (e.g. CAS, Name, ...) to InChIKey, the DoOR default identifier.
}
\examples{
getKey("123-92-2")
getKey("isopentyl acetate", "Name")
getKey("C(C(C)C)COC(=O)C", "SMILES")

ALimage(getKey("123-92-2"))
}

