% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getNormalizedResponses.R
\name{getNormalizedResponses}
\alias{getNormalizedResponses}
\title{Find normalised receptor responses}
\usage{
getNormalizedResponses(odors, zero = default.val("zero"),
  response_matrix = default.val("response.matrix"))
}
\arguments{
\item{odors}{character vector; one or more odors provided as InChIKey.}

\item{zero}{InChIKey of background that should be set to zero. The default is "SFR", i.e. the spontaneous
firing rate.}

\item{response_matrix}{a data frame; as e.g. "response.matrix" that is loaded
by \code{\link{modelRP}}. It is also possible to create this frame manually
using \code{\link{modelRP}}.}
}
\description{
given a chemical, get normalised receptor responses from all studies in the
database.
}
\examples{
library(DoOR.data)
odors <- c("MLFHJEHSLIIPHL-UHFFFAOYSA-N","OBNCKNCVKJNDBV-UHFFFAOYSA-N","IKHGUXGNUITLKF-UHFFFAOYSA-N")
data(response.matrix)
result <- getNormalizedResponses(odors, response_matrix = response.matrix)
}
\seealso{
\code{\link{modelRP}},\code{\link{CreateDatabase}}
}
\keyword{data}

