% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResponses.R
\name{getResponses}
\alias{getResponses}
\title{Find receptor responses}
\usage{
getResponses(odorant, responseRange = default.val("response.range"),
  Or.list = load2list())
}
\arguments{
\item{odorant}{a single odor provided as InChIKey}

\item{responseRange}{data frame, response ranges of studies}

\item{Or.list}{a list contains reponse data of all available receptors. It
can be loaded using \code{\link{load2list}}.}
}
\description{
given a chemical, get original receptor responses from all studies in the
database.
}
\details{
output is a data frame containing response values of given odor across
receptors from all available studies.
}
\examples{

library(DoOR.data)
loadData()
responses <- getResponses(odor = 'MLFHJEHSLIIPHL-UHFFFAOYSA-N')

}
\keyword{data}

