% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/inverse_modelfunction_linear.R
\name{inverse_modelfunction_linear}
\alias{inverse_modelfunction_linear}
\title{inverse Linear Model}
\usage{
inverse_modelfunction_linear(input, parms)
}
\arguments{
\item{input}{numeric vector of data values}

\item{parms}{numeric vector; parameters with given names: "Intercept",
"Slope"}
}
\description{
inverse Linear Model
}
\details{
expression : (input - Intercept) / Slope
}
\examples{
x <- seq(0.1,1,length=20)
y <- inverse_modelfunction_linear(input = x, parms = c(Intercept = 0.1, Slope = 0.2) )
plot(x,y)
}
\keyword{math}

