% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inverse_modelfunction_sigmoid.R
\name{inverse_modelfunction_sigmoid}
\alias{inverse_modelfunction_sigmoid}
\title{inverse Logistic Model}
\usage{
inverse_modelfunction_sigmoid(input, parms)
}
\arguments{
\item{input}{numeric vector of data values}

\item{parms}{numeric vector; parameters with given names: "Asym", "xmid",
"scal"}
}
\description{
inverse Logistic Model
}
\details{
expression : xmid - scal * log((Asym / input) - 1)
}
\examples{
x <- seq(0.1,1,length=20)
y <- inverse_modelfunction_sigmoid(input = x, parms = c(Asym = 2, xmid = 0.2, scal = 0.3) )
plot(x,y)
}
\references{
Jose Pinheiro and Douglas Bates
}
\seealso{
\code{\link{SSlogis}}
}
\keyword{math}

