% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loadRD.R
\name{loadRD}
\alias{loadRD}
\title{Load Data}
\usage{
loadRD(odorantReceptors = default.val("ORs"), supportData = TRUE)
}
\arguments{
\item{odorantReceptors}{data frame; containing receptor or ORN names and
their expression.}

\item{supportData}{logical; if TRUE, load support data.}
}
\description{
Load all datasets including the precomputed response model.
}
\details{
Please load ORs from data package DoOR.data by typing (\code{data(ORs)})
before use.
}
\examples{
library(DoOR.data)
library(DoOR.function)
loadRD()
}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>
}
\keyword{data}

