% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_receptor.R
\name{map_receptor}
\alias{mapReceptor}
\alias{map_receptor}
\title{map_receptor}
\usage{
map_receptor(data,
  response_matrix = door_default_values("door_response_matrix"), sub,
  threshold.p, threshold.cor, nshow)
}
\arguments{
\item{data}{data frame, containing two columns, one called "odorants" and one
"responses" providing InChIKeys and odorant responses respectively.}

\item{response_matrix}{output is a numeric vector that contains the Pearson
Correlation Coefficient between given data and selected consensus data in}

\item{sub}{character, a subset of responding units
returned response matrix}

\item{threshold.p}{numeric, a p-value threshold, only correlations below will be returned}

\item{threshold.cor}{numeric, a correlation-coefficient threshold, only correlations above will be returned}

\item{nshow}{numeric, if defined, only this number of results will be}
}
\description{
Identifying the source of unknown response data by correlating it agains all
DoOR responding units.
}
\examples{
# load data
load_door_data(nointeraction = TRUE)

# pick example data
data <- data.frame(odorants  = Or22a$InChIKey,
                   responses = Or22a$Hallem.2004.EN)
data <- na.omit(data)

# find the corresponding receptor / responding unit
map_receptor(data = data)
}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>

Daniel Münch <\email{daniel.muench@uni-konstanz.de}>
}

