% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_response.R
\name{model_response}
\alias{modelRP}
\alias{model_response}
\title{Generates a model response}
\usage{
model_response(da, select.MDValue = door_default_values("select.MDValue"),
  overlapValues = door_default_values("overlapValues"),
  responseRange = door_default_values("door_response_range"),
  weightGlobNorm = door_default_values("weight.globNorm"),
  glob.normalization = door_default_values("glob.normalization"),
  plot = door_default_values("plot"))
}
\arguments{
\item{da}{data frame, a selected receptor containing measured responses from
studies.}

\item{select.MDValue}{numeric, threshold on the MD for rejecting a fit.}

\item{overlapValues}{numeric, a criterion using to refuse a data set that
has not enough overlap value.}

\item{responseRange}{data frame, contains response ranges for all studies.}

\item{weightGlobNorm}{data frame, a binary data matrix, 1 indicates given
odor has been measured in given study, NA indicates NOT.}

\item{glob.normalization}{logical, default is \code{TRUE}, performs a global
normalization for the model response. Otherwise (\code{FALSE}) response
values will be given in value from 0 to 1.}

\item{plot}{logical, If \code{FALSE}, plotting is suppressed. Default is
\code{FALSE}.}
}
\description{
Runs the DoOR algorithm, that merges all measurements for one receptor into
a common response model.
}
\details{
Merging a data is processed by following: \enumerate{ \item Normalize all
response data in value [0,1].  \item Compute the correlation between studies
and selected the best pair using \code{\link{select_model}}.  \item Merge the
first pair using function \code{\link{project_points}}.  \item Add other
datasets if the correlation between the growing model response and the new
dataset is below the correlation threshold (select.MDValue). Datasets
excluded based on this criterion will be appended in a separate list.  }
}
\examples{

library(DoOR.data)
data(Or35a)
data(weight.globNorm)
data(door_response_range)
RP.Or35a <- model_response(Or35a, plot = TRUE)

}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>
}
\keyword{data}

