% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modelfunction.R
\name{modelfunction}
\alias{modelfunction}
\title{estimate the parameters for all optional model function}
\usage{
modelfunction(x, y)
}
\arguments{
\item{x,y}{data vectors; in which to evaluate the parameters in given
formula.}
}
\description{
estimate the parameters for all optional model function
}
\details{
This function is used to estimate the parameters for optional model
functions with lm() for linear model and nls() for nonlinear function:
linear (\code{\link{modelfunction_linear}}), exponential
(\code{\link{modelfunction_exp}}), sigmoid (\code{\link{SSlogis}}),
asymptotic model with x intercept (\code{\link{SSasympOff}}), asympototic
model with y intercept (\code{\link{SSasymp}}) and their inverse function,
\code{\link{inverse_modelfunction_linear}},
\code{\link{inverse_modelfunction_exp}},
\code{\link{inverse_modelfunction_sigmoid}},
\code{\link{inverse_modelfunction_asympOff}},
\code{\link{inverse_modelfunction_asymp}}. The output is a list containing
estimated parameter, function and inverse function expression, applied range
for middle function, Leibniz's notation for computing curvic length and mean
distance value "MD" of all optional model function.
}
\examples{
library(DoOR.data)
loadRD()
x<-DoORnorm(Or23a[,'Hallem.2004.EN'])
y<-DoORnorm(Or23a[,'Hallem.2006.EN'])
MF_xy <- modelfunction(x=x,y=y)
}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>
}
\keyword{math}

