% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelfunction_asymp_inverse.R
\name{modelfunction_asymp_inverse}
\alias{modelfunction_asymp_inverse}
\title{inverse Asymptotic Model}
\usage{
modelfunction_asymp_inverse(input, parms)
}
\arguments{
\item{input}{numeric vector of data values}

\item{parms}{numeric vector; parameters with given names: "Asym", "R0",
"lrc"}
}
\description{
inverse Asymptotic Model
}
\details{
expression : -log((input-Asym)/(R0-Asym))/exp(lrc)
}
\examples{
x <- seq(0.1,1,length=20)
y <- modelfunction_asymp_inverse(input = x, parms = c(Asym = 2, R0 = 0.2, lrc = 0.3) )
plot(x,y)
}
\references{
Jose Pinheiro and Douglas Bates
}
\keyword{math}

