% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelfunction_exp.R
\name{modelfunction_exp}
\alias{modelfunction_exp}
\title{exponential Model}
\usage{
modelfunction_exp(input, parms)
}
\arguments{
\item{input}{numeric vector of data values}

\item{parms}{numeric vector; parameters with given names: "a", "b", "c"}
}
\description{
exponential Model
}
\details{
expression : a + b * exp (c * input)
}
\examples{
x <- rnorm(20)
y <- modelfunction_exp(input = x, parms = c(a = 0.1, b = 0.2, c = 0.3) )
plot(x,y)
}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>
}
\seealso{
\code{\link{SSexpo}}
}
\keyword{math}

