% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelfunction_exp_inverse.R
\name{modelfunction_exp_inverse}
\alias{modelfunction_exp_inverse}
\title{inverse exponential Model}
\usage{
modelfunction_exp_inverse(input, parms)
}
\arguments{
\item{input}{numeric vector of data value}

\item{parms}{numeric vector; parameters with given names: "a", "b", "c"}
}
\description{
inverse exponential Model
}
\details{
expression : (log((input-a)/b))/c
}
\examples{
x <- seq(0.1,1,length=20)
y <- modelfunction_exp_inverse(input = x, parms = c(a = 0, b = 0.2, c = 0.3) )
plot(x,y)
}
\seealso{
\code{\link{SSexpo}}
}
\keyword{math}

