% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelfunction_linear.R
\name{modelfunction_linear}
\alias{modelfunction_linear}
\title{Linear Model}
\usage{
modelfunction_linear(input, parms)
}
\arguments{
\item{input}{numeric vector of data values}

\item{parms}{numeric vector; parameters with given names: "Intercept",
"Slope"}
}
\description{
Linear Model
}
\details{
expression : Intercept + Slope * input
}
\examples{
x <- rnorm(20)
y <- modelfunction_linear(input = x, parms = c(Intercept = 0.1, Slope = 0.2) )
plot(x,y)
}
\keyword{math}

