% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelfunction_linear_inverse.R
\name{modelfunction_linear_inverse}
\alias{modelfunction_linear_inverse}
\title{inverse Linear Model}
\usage{
modelfunction_linear_inverse(input, parms)
}
\arguments{
\item{input}{numeric vector of data values}

\item{parms}{numeric vector; parameters with given names: "Intercept",
"Slope"}
}
\description{
inverse Linear Model
}
\details{
expression : (input - Intercept) / Slope
}
\examples{
x <- seq(0.1,1,length=20)
y <- modelfunction_linear_inverse(input = x, parms = c(Intercept = 0.1, Slope = 0.2) )
plot(x,y)
}
\keyword{math}

