% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelfunction_sigmoid_inverse.R
\name{modelfunction_sigmoid_inverse}
\alias{modelfunction_sigmoid_inverse}
\title{inverse Logistic Model}
\usage{
modelfunction_sigmoid_inverse(input, parms)
}
\arguments{
\item{input}{numeric vector of data values}

\item{parms}{numeric vector; parameters with given names: "Asym", "xmid",
"scal"}
}
\description{
inverse Logistic Model
}
\details{
expression : xmid - scal * log((Asym / input) - 1)
}
\examples{
x <- seq(0.1,1,length=20)
y <- modelfunction_sigmoid_inverse(input = x, parms = c(Asym = 2, xmid = 0.2, scal = 0.3) )
plot(x,y)
}
\references{
Jose Pinheiro and Douglas Bates
}
\seealso{
\code{\link{SSlogis}}
}
\keyword{math}

