% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelRPSEQ.R
\name{modelRPSEQ}
\alias{modelRPSEQ}
\title{Generates a model response and merge data in given sequence}
\usage{
modelRPSEQ(data, SEQ, overlapValues = default.val("overlapValues"),
  plot = F)
}
\arguments{
\item{data}{data frame; odorant response data, e.g. Or22a.}

\item{SEQ}{character vector; containing the names of studies indicating
given sequence for merging data.}

\item{overlapValues}{minimum overlap between studies to perfom a merge}
}
\description{
Generates a model response and merge data in given sequence.
}
\details{
# modelRPSEQ.R:
#################

# merges studies in a given sequence (determined by the user or by exhaustive enumeration and choosing the optimal sequence)


# input parameters:
####################


# data  : data frame; odorant response data for a given receptor, e.g. Or22a
# SEQ 	: character vector; contains the names of studies that measured this receptor in a specific order (the merging sequence)

# output is a numeric vector: response values
}
\examples{
library(DoOR.data)
data(Or35a)
data(response.range)
SEQ <- c("Hallem.2006.EN","Kreher.2008.EN","Yao.2005.WT")
selected.merg <- modelRPSEQ(Or35a, SEQ = SEQ)
}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>
}
\keyword{data}

