% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pyramid.R
\name{orderPyramid}
\alias{orderPyramid}
\title{orderPyramid}
\usage{
orderPyramid(x, na.rm = F)
}
\arguments{
\item{x}{vector that should be ordered}

\item{na.rm}{whether ornot to remove missing values}
}
\value{
integer, the "pyramidal" order of x
}
\description{
returns the "pyramidal" order for a vector with numeric values
}
\examples{
data  <- rnorm(n = 20, mean = 2, sd = .2)
order.pyramid <- orderPyramid(data)
plot(data)
plot(data[order.pyramid])
}

