% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permutations.R
\name{permutations}
\alias{permutations}
\title{calculate all possible permutations}
\usage{
permutations(n)
}
\arguments{
\item{n}{length of vector to permute}
}
\description{
calculate all possible permutations
}

