% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotReceptors.R
\name{PlotReceptors}
\alias{PlotReceptors}
\title{Bar Plot of Responses of given receptor}
\usage{
PlotReceptors(data, odor.data, tag = default.val("tag"), name.title,
  pic.height, file.name, cex.title = 1, x.range, cex.axes = 1,
  cex.tag = 1, PDF = FALSE, JPEG = FALSE, pic.width = 864,
  point.size = 17, col.extrem = default.val("col.extrem"))
}
\arguments{
\item{data}{data frame; containing the names of odorant receptors, CAS
number(s) and normalized response values.}

\item{odor.data}{data frame; chemical data of odors.
\code{\link[DoOR.data]{odor}}}

\item{tag}{character string; default tag is the CAS number.}

\item{name.title}{character string; the main title of plot. Dafaults to
"Chemical Responses Crossing Odor Receptors".}

\item{pic.height}{numeric; the height of the device; the default is 1296.}

\item{file.name}{character string; file name; if missing, the default file
name is "PlotReceptors.pdf" or "PlotReceptors.jpeg"}

\item{cex.title}{numeric; the magnification to be used for main titles. The
default is 1.}

\item{x.range}{numeric vectors of length 2; giving the x coordinates ranges.}

\item{cex.axes}{numeric; the magnification to be used for axes. The default
is 1.}

\item{cex.tag}{numeric; the magnification to be used for tag. The default is
1.}

\item{PDF}{logical; if \code{TRUE}, then the bar plot will be saved as PDF
file. The default is \code{FALSE}.}

\item{JPEG}{logical; if \code{TRUE}, then the bar plot will be saved as JPEG
file. The default is \code{FALSE}.}

\item{pic.width}{numeric; the width of the device; the default is 864.}

\item{point.size}{numeric; pointsize of plotted text; one point is
approximately one pixel, and the default is 17.}

\item{col.extrem}{a two characters vector; coding the colour of each end.
The default is c("blue", "red").}
}
\description{
Create a horizontal bar plot and show the response profile of all commented
odor receptors to odor(s).
}
\details{
This function is used to visualize the response profile of all commented
odor receptors to odors. For \code{PlotReceptors}, the given data should be
generated by \code{\link{findRespNorm}} or the columns of data must be name
of odor receptors, CAS number and normalized response value. The color bar
from green to magenta indicates that the response intensity from most
inhibitory to most excitatory. White area between green and magenta
indicates that the odor response is weak.
}
\examples{
library(DoOR.data)
loadRD()
data(odor)
cas<-c("71-36-3","123-92-2","79-09-4")
result2<-findRespNorm(cas,responseMatrix = response.matrix)
PlotReceptors(result2,odor.data=odor,tag="Name",PDF=TRUE)
}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>
}
\seealso{
\code{\link{findRespNorm}}
}
\keyword{iplot}

