% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_odorant.R
\name{private_odorant}
\alias{privateOdorant}
\alias{private_odorant}
\title{private_odorant}
\usage{
private_odorant(receptor, sensillum = FALSE,
  response_matrix = door_default_values("door_response_matrix"),
  door_mappings = door_default_values("door_mappings"),
  zero = door_default_values("zero"), nshow = 5, tag)
}
\arguments{
\item{receptor}{character, name of a DoOR responding unit (one of \code{ORs$Or})}

\item{sensillum}{logical, restrict the search to the sensillum the receptor is expressed in?}

\item{response_matrix}{DoOR response matrix, the input data to perform the search on}

\item{door_mappings}{the data frame containing the mapping information}

\item{zero}{character, an odorant that should be set to 0}

\item{nshow}{numeric, the number of private odorants to return}

\item{tag}{character, the chemical identifier to give the odorant names in (on of \code{colnames(odor)})}
}
\value{
a data.frame containing odorants and the response in the receptor of interest as well as the maximum response of the remaining receptors and their difference
}
\description{
return an odorant that activates the receptor of interest exclusively
}
\examples{
library(DoOR.data)
private_odorant("Gr21a.Gr63a", tag = "Name")
private_odorant("Or22a", tag = "Name", sensillum = TRUE)

}

