% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pyramid.R
\name{pyramid}
\alias{pyramid}
\title{pyramid}
\usage{
pyramid(n)
}
\arguments{
\item{n}{length of the sequence to be returned}
}
\value{
a "pyramidally ordered" sequence of integers with length n
}
\description{
returns a "pyramidal" sequence with the length supplied by n
}
\examples{
pyramid(10)
}

