% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_study.R
\name{remove_study}
\alias{removeStudy}
\alias{remove_study}
\title{Remove a study from DoOR}
\usage{
remove_study(study, receptors = door_default_values("ORs"),
  responseRange = door_default_values("door_response_range"),
  weightGlobNorm = door_default_values("door_global_normalization_weights"))
}
\arguments{
\item{study}{a string containing the name of the study you want to remove
(e.g. 'Bruyne.2001.WT')}

\item{receptors}{a vector of all the receptors to be checked. Defaults to all
receptors exidting in DoOR.}

\item{responseRange}{the dataframe containing the info about the response
ranges of all studies (\code{door_response_range})}

\item{weightGlobNorm}{the dataframe containing the info about the relative
weights between receptors (\code{door_global_normalization_weights})}
}
\description{
Use this function to remove a study from the DoOR database.
\code{import_new_data.R} uses this function when it detects an existing study
during the import process (e.g. because you imported updated data).
}
\examples{
library(DoOR.data)
load_door_data()
remove_study('Bruyne.2001.WT')

}
\author{
Daniel Münch <\email{daniel.muench@uni-konstanz.de}>
}
\seealso{
\code{\link{import_new_data}}
}

