% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_sfr.R
\name{reset_sfr}
<<<<<<< HEAD:man/reset_sfr.Rd
\alias{resetSFR}
=======
>>>>>>> 9d73a9bc3fe04986e1b96d7d4c8b786026637987:man/resetSFR.Rd
\alias{reset_sfr}
\title{reset SFR}
\usage{
reset_sfr(x, sfr)
}
\arguments{
\item{x}{numeric or DoOR response matrix, input values}

\item{sfr}{numeric or character, either a value to subtract if x is a vector or an InChIKey if x is a DoOR response matrix}
}
\description{
A function for reseting SFR to zero
}
\details{
Performs a simple subtraction of the SFR value.
}
\examples{

library(DoOR.data)
data(response.matrix)
response.matrix.SFRreset <- reset_sfr(response.matrix, "SFR")

}
\author{
Daniel Münch <\email{daniel.muench@uni-konstanz.de}>
}

