% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/resetSFR_old.R
\name{resetSFR_old}
\alias{resetSFR_old}
\title{reset SFR}
\usage{
resetSFR_old(x, sfr)
}
\arguments{
\item{x}{a numeric vector of input values.}

\item{sfr}{a numieric parameter; indicating spontaneous firing rate value.}
}
\description{
A function for reseting SFR to zero
}
\details{
The expression is \code{(x-sfr)*(max(x))/(max(x)-sfr)}. If the spontaneous
firing rate is missing \code{(NA)}, the value of sfr is given 0.
}
\examples{
library(DoOR.data)
data(response.matrix)
x=response.matrix[,'Or22a']
y=resetSFR(x=x,sfr=x[1])
}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>
}
\keyword{math}

