% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_model.R
\name{select_model}
\alias{selectModel}
\alias{select_model}
\title{compute the data pairwise and and selects a pair with the lowest "MD" value.}
\usage{
select_model(candidate, data_candidate, merged_data,
  overlapValues = door_default_values("overlapValues"),
  merged = door_default_values("merged"))
}
\arguments{
\item{candidate}{a character vector, contains the names of studies.}

\item{data_candidate}{a data frame, odorant response data that only contains
value columns.}

\item{merged_data}{numeric vector, merged data}

\item{overlapValues}{numeric, a criterion using to refuse a data set that
has not enough overlap value.}

\item{merged}{logical, if merged is \code{TRUE}, calculate models between
merged_data and candidate data. If \code{FALSE}, calculate models between
candidates.}
}
\description{
compute the data pairwise using function \code{\link{calculate_model}} and selects
a pair with the lowest "MD" value.
}
\details{
This function is used in \code{\link{model_response}} to select the first pair or
next data set for merging. The output is a list containing "selected.x" and
"selected.y" that specify which data plots against another, and "best.model"
that is used in function \code{\link{project_points}}.
}
\examples{

library(DoOR.data)
data(ac3B)
studies<-names(ac3B)[c(7:9)]
data_candidate<-ac3B[,c(7:9)]
Norm_data_candidate <-apply(data_candidate, 2, door_norm)
select_model(candidate = studies, data_candidate = Norm_data_candidate, merged = FALSE)

}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>
}
\seealso{
\code{\link{project_points}},\code{\link{model_response}}
}
\keyword{data}

