% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/selectModel.R
\name{selectModel}
\alias{selectModel}
\title{compute the data pairwise and and selects a pair with the lowest "MD" value.}
\usage{
selectModel(candidate, data_candidate, merged_data,
  overlapValues = default.val("overlapValues"),
  merged = default.val("merged"))
}
\arguments{
\item{candidate}{a character vector, contains the names of studies.}

\item{data_candidate}{a data frame, odorant response data that only contains
value columns.}

\item{merged_data}{numeric vector, merged data}

\item{overlapValues}{numeric, a criterion using to refuse a data set that
has not enough overlap value.}

\item{merged}{logical, if merged is \code{TRUE}, calculate models between
merged_data and candidate data. If \code{FALSE}, calculate models between
candidates.}
}
\description{
compute the data pairwise using function \code{\link{calModel}} and selects
a pair with the lowest "MD" value.
}
\details{
This function is used in \code{\link{modelRP}} to select the first pair or
next data set for merging. The output is a list containing "selected.x" and
"selected.y" that specify which data plots against another, and "best.model"
that is used in function \code{\link{projectPoints}}.
}
\examples{
library(DoOR.data)
data(ac3B)
studies<-names(ac3B)[c(7:9)]
data_candidate<-ac3B[,c(7:9)]
Norm_data_candidate <-apply(data_candidate, 2, DoORnorm)
selectModel(candidate = studies, data_candidate = Norm_data_candidate, merged = FALSE)
}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>
}
\seealso{
\code{\link{projectPoints}},\code{\link{modelRP}}
}
\keyword{data}

