% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sparse.R
\name{sparse}
\alias{sparse}
\title{Calculate the sparseness of a vector.}
\usage{
sparse(x, method = "ltk")
}
\arguments{
\item{x}{numerical input vector}

\item{method}{type of sparseness measure, either 'ltk' for lifetime kurtosis
or 'lts' for lifetime sparseness (see references).}
}
\description{
Sparseness can be calculated as lifetime kurtosis (LTK, Willmore and
Tolhurst, 2001) or as lifetime sparseness (LTS, Bhandawat et al., 2007).
}
\details{
LTS is scales between \[0,1\] while LTK is not restricted. LTS only
  takes positive values.
}
\author{
Daniel Münch <\email{daniel.muench@uni-konstanz.de}>
}
\references{
Bhandawat, V., Olsen, S.R., Gouwens, N.W., Schlief, M.L., Wilson,
  R.I., 2007. Sensory processing in the Drosophila antennal lobe increases
  reliability and separability of ensemble odor representations. Nature
  neuroscience 10, 1474–82. doi:10.1038/nn1976

Willmore, B., Tolhurst, D.J., 2001. Characterizing the sparseness
  of neural codes. Network 12, 255–270. doi:10.1080/net.12.3.255.270
}
\keyword{kurtosis}
\keyword{sparseness,}

