% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_id.R
\name{trans_id}
<<<<<<< HEAD:man/trans_id.Rd
\alias{transID}
=======
>>>>>>> 9d73a9bc3fe04986e1b96d7d4c8b786026637987:man/transID.Rd
\alias{trans_id}
\title{trans_id}
\usage{
trans_id(x, from = "CAS", to = "InChIKey",
  odor_data = door_default_values("odor"))
}
\arguments{
\item{x}{character vector, one or many chemical identifiers}

\item{from}{character, the type of identifier to translate from (one of the column names of ``odor``)}

\item{to}{character, the type of identifier to translate from (one of the column names of ``odor``)}

\item{odor_data}{the data frame containing the odor information (defaults to ``odor``).}
}
\value{
character vector of translated chemical identifiers
}
\description{
Translate chemical identifiers from one to the other.
}
\examples{
library(DoOR.data)
trans_id("123-92-2")
trans_id("isopentyl acetate", "Name")
trans_id("C(C(C)C)COC(=O)C", "SMILES", "Name")

}

