% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transID.R
\name{transID}
\alias{transID}
\title{transID}
\usage{
transID(x, from = "CAS", to = "InChIKey",
  odor_data = default.val("odor_data"))
}
\arguments{
\item{x}{character vector; one or many chemical identifiers}

\item{from}{character; the type of identifier to translate from (one of the column names of ``odor``)}

\item{to}{character; the type of identifier to translate from (one of the column names of ``odor``)}

\item{odor_data}{the data frame containing the odor information (defaults to ``odor``).}
}
\value{
character vector of translated chemical identifiers
}
\description{
Translate chemical identifiers from one to the other.
}
\examples{
transID("123-92-2")
transID("isopentyl acetate", "Name")
transID("C(C(C)C)COC(=O)C", "SMILES", "Name")

DoORplot_ALmap(getKey("123-92-2"))
}

