% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_door_database.R
\name{update_door_database}
\alias{update_door_database}
\title{update response matrix}
\usage{
update_door_database(receptor, permutation = TRUE, perm,
  response_matrix_nn = door_default_values("door_response_matrix_non_normalized"),
  response_matrix = door_default_values("door_response_matrix"),
  responseRange = door_default_values("door_response_range"),
  weightGlobNorm = door_default_values("door_global_normalization_weights"),
  select.MDValue = door_default_values("select.MDValue"), strict = TRUE,
  overlapValues = door_default_values("overlapValues"),
  door_excluded_data = door_default_values("door_excluded_data"),
  plot = FALSE)
}
\arguments{
\item{receptor}{character string, name of given odorant receptor.}

\item{permutation}{logical, if TRUE, the sequence is chosen from permutation,
if FALSE, sequence is chosen by the routine process.}

\item{perm}{a matrix with one sequence of study names per row, if empty, all
possible permutations of study names will be provided.}

\item{response_matrix_nn}{data frame, response data that has not been
globally normalized.}

\item{response_matrix}{data frame, globally normalized response data.}

\item{responseRange}{data frame, response range of studies.}

\item{weightGlobNorm}{data frame, weight matrix for global normalization.}

\item{select.MDValue}{the minimum mean distance between studies to perfom a
merge (used if permutation == FALSE or if permutation == TRUE AND strict ==
TRUE)}

\item{strict}{logical, if TRUE merging a permutation will be stopped once a
single merge has a mean distance above select.MDValue (only valid if
permutation == TRUE)}

\item{overlapValues}{minimum overlap between studies to perfom a merge}

\item{door_excluded_data}{the data frame that contains the list of excluded data
sets.}

\item{plot}{logical}
}
\description{
update the globally \code{response matrix} and the unglobally normalized
response matrix \code{door_response_matrix_non_normalized} by introducing new
consensus response data of given receptor.
}
\details{
The merging sequence could be arranged by the routine process (using
\code{\link{model_response}} or taking the optimized sequence that is chosen from
permutations. The mean correlation between merged responses and each original
recording will be computed for each permutation, the optimozed sequence is
with the highest correlation.
}
\examples{
\dontrun{
# load data
library(DoOR.data)
load_door_data()
# update the entry "Or67b" of data "door_response_matrix" and
# "door_response_matrix_non_normalized" with permutations.
 update_door_database(receptor="Or67b", permutation = TRUE)
}
}
\seealso{
\code{\link{model_response}},\code{\link{model_response_seq}}
}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>

Shouwen Ma <\email{daniel.muench@uni-konstanz.de}>
}
\keyword{data}
