% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_door_odorinfo.R
\name{update_door_odorinfo}
\alias{update_door_odorinfo}
\alias{updateOdorInfo}
\title{update_door_odorinfo}
\usage{
update_door_odorinfo()
}
\description{
Update the DoOR odor data with info from \code{odor}. For the function to
work, all DoOR data has to be loaded to the current environment.
}
\examples{
# load data
load_door_data(nointeraction = TRUE)  
# modify odor
odor[1,1] <- "acid"

# run 
update_door_odorinfo()

# check that data sets have been updated
head(Or22a)


}
\author{
Daniel Münch, \email{daniel@muench.bio}
}
