% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_odorinfo.R
\name{update_odorinfo}
\alias{updateOdorInfo}
\alias{update_odorinfo}
\title{update_odorinfo}
\usage{
update_odorinfo()
}
\description{
Update the DoOR odor data with info from \code{odor}. For the function to
work, all DoOR data has to be loaded.
}
\author{
Daniel Münch, \email{daniel@muench.bio}
}

