% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/updateOdorInfo.R
\name{updateOdorInfo}
\alias{updateOdorInfo}
\title{Update the DoOR odor data with info from \code{odor}}
\usage{
updateOdorInfo()
}
\description{
Update the DoOR odor data with info from \code{odor}
}
\author{
Daniel Münch, \email{daniel@muench.bio}
}

