% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/version.function.R
\name{version.function}
\alias{version.function}
\title{DoOR.functions package version}
\usage{
version.function()
}
\description{
Display the DoOR.functions package version when loading.
}
\keyword{DoOR}

