% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/backProject.R
\name{backProject}
\alias{backProject}
\title{backProject}
\usage{
backProject(template.data, responding.unit,
  response_matrix = default.val("response.matrix"))
}
\arguments{
\item{template.data}{data frame, the data that provides the scale to
transform to, containing InChIKeys in a column called "odorants" and
responses in a column called "responses"}

\item{responding.unit}{character, the name of the receptor/OSN/glomerulus
which responses should be transformed}

\item{response_matrix}{DoOR response mnatrix, the source data is picked from
  here}
}
\value{
Output of backProject is a list containing a data frame with the
  backprojected data, the original data, the data used as a template and the
  data that resulted from fitting source and template (before rescaling to
  the template scale). additionaly the parameters of the linear fit between
  the source and template response scale is returned.
}
\description{
project the model response values back into your scale of interest (spike,
deltaF/F...)
}
\details{
The process of back projection is the following: \itemize{ \item 1. rescale
both data sets to [0,1], \item 2. find the best fitting model between
"bp.data" and "cons.data" (lowest MD value), \item 3. project the consensus
data onto the fitted curv, these are now our normalized, back projected
responses \item 4. rescale all responses to the scale of the original data
via a linear fit.  }
}
\examples{
data(Or22a)
data(response.matrix)
template.data <- data.frame(odorants = Or22a$InChIKey,
                            responses = Or22a$Hallem.2004.EN)
bp <- backProject(template.data, "Or22a")

plot(bp$backprojected$original.data,
     bp$backprojected$backprojected.data,
     xlab = "DoOR consensus response",
     ylab = "backprojected data [spikes, Hallem.2004.EN]"
)
}
\author{
Daniel Münch <\email{daniel.muench@uni-konstanz.de}>

Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>
}

