% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calModel.R
\name{calModel}
\alias{calModel}
\title{select the best model function}
\usage{
calModel(x, y, select.MD = default.val("select.MD"))
}
\arguments{
\item{x,y}{data vectors from study x and y (can contain NA)}

\item{select.MD}{logical; if TRUE, only the best model function (in terms of
MD) will be returned.}
}
\description{
\code{calModel} is used to return the best model function that represent the
relationship between responses from study x and y..
}
\details{
\code{calModel} chooses the best model function from following: linear
(\code{\link{modelfunction_linear}}), exponential function
(\code{\link{modelfunction_exp}}), sigmoid
(\code{\link{modelfunction_sigmoid}}), asymptotic model with x intercept
(\code{\link{modelfunction_asympOff}}), asympototic model with y intercept
(\code{\link{modelfunction_asymp}}) and their inverse model
functions:(\code{\link{inverse_modelfunction_linear}}),
(\code{\link{inverse_modelfunction_exp}}),
(\code{\link{inverse_modelfunction_sigmoid}}),
(\code{\link{inverse_modelfunction_asympOff}}) and
(\code{\link{inverse_modelfunction_asymp}}).
}
\examples{
library(DoOR.data)
data(Or35a)
x <- DoORnorm(Or35a[,5])
y <- DoORnorm(Or35a[,8])
calM_xy <- calModel(x, y, select.MD = default.val("select.MD"))
}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>
}
\seealso{
\code{\link{SSlogis}},\code{\link{SSasympOff}},
\code{\link{SSasymp}}
}
\keyword{data}

