% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CreateDatabase.R
\name{CreateDatabase}
\alias{CreateDatabase}
\title{Compose a Response Matrix of All Odor Receptors}
\usage{
CreateDatabase(tag = default.val("tag"),
  select.MDValue = default.val("select.MDValue"),
  overlapValues = default.val("overlapValues"), glob.normalization = T, ...)
}
\arguments{
\item{tag}{character string; format for rownames; possibilities: tag="CAS",
tag="CID", tag="Name"}

\item{select.MDValue}{a numeric; threshold on the MD, this is used to reject
studies that do not align sufficiently well to the response model}

\item{overlapValues}{numeric; a criterion using to refuse a data set that
has not enough overlap value.}

\item{...}{pass more parameters to \code{\link{modelRP}} (e.g. \code{glob.normalization == F} to build a non-normalized response matrix)}
}
\description{
computes the complete response model for all receptors in the database
(calls \code{\link{modelRP}} for all receptors)
}
\examples{
library(DoOR.data)
loadRD()
# mydatabase <- CreateDatabase()
}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>

Shouwen Ma <\email{daniel.muench@uni-konstanz.de}>
}
\seealso{
\code{\link{modelRP}}
}
\keyword{data}

