% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOREst.R
\name{DoOREst}
\alias{DoOREst}
\title{Estimate the missing entries in a response data}
\usage{
DoOREst(da, nodor, method = "PC")
}
\arguments{
\item{da}{a data frame or matrix; contaning the consensus response values.}

\item{nodor}{a numeric value; specifying the number of the selected odors.}

\item{method}{character string; specifying the method ("PC" (Pearson's
coefficient) and "Knn" (k nearest neighbors)) for estimation, the default is
"PC".}
}
\description{
Estimate the missing entries in a response data
}
\details{
A warp programe using Pearson Correlation or k-nearest neighbors
estimate the missing entries in a response data.
}
\examples{
library(DoOR.data)
data(response.matrix)
# est_data <- DoOREst(da=response.matrix, nodor = 3)
}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>
}
\references{
Kim, H.; Golub, G. H. & Park, H., Missing value estimation for
DNA microarray gene expression data: local least squares imputation., 2005,
Bioinformatics, 21, 187-198
}
\keyword{math}

