% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoORplot_acrossOSNs.R
\name{DoORplot_acrossOSNs}
\alias{DoORplot_acrossOSNs}
\title{DoORplot_acrossOSNs}
\usage{
DoORplot_acrossOSNs(odorants, responseMatrix = default.val("response.matrix"),
  zero = default.val("zero"), tag = "Name", plot.type = 1)
}
\arguments{
\item{odorants}{character vector; one or several InChIKeys}

\item{responseMatrix}{DOOR response matrix; contains the data to plot}

\item{zero}{character; InChIKey of the odorant that should be set to 0 (e.g. SFR)}

\item{tag}{character; the chemical identifier to use in the plot, one of \code{colnames(odor)}}

\item{plot.type}{interger; 1 or 2, defining the type of plot (1: facet_grid of odorants * sensillae; 2: facet_wrap across OSNs)}
}
\value{
a ggplot2 object
}
\description{
plot the activation patterns of one or several odorants across OSNs
}
\details{
DoOR response profiles will be selected and ordered according to the OSNs they are related to. Several DoOR response profiles might exist for a given OSN (e.g. one for the OSN itself and one for the OSNs misexpressed receptor protein) but only one will be shown. Which DoOR profile is mapped to which OSN is controlled via the "code.OSN" column in \code{DoORmapings}.
}
\examples{
odorants <- sapply(c("1-butanol", "isopentyl acetate", "carbon dioxide", "water"), function(x) getKey(x, type = "Name"))
DoORplot_acrossOSNs(odorants)
DoORplot_acrossOSNs(odorants, plot.type = 2)
}
\author{
Daniel Münch <\email{daniel.muench@uni-konstanz.de}>
}

