% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoORplot_acrossReceptors.R
\name{DoORplot_acrossReceptors}
\alias{DoORplot_acrossReceptors}
\title{DoORplot_acrossReceptors}
\usage{
DoORplot_acrossReceptors(odorant,
  response_matrix = default.val("response.matrix"),
  odor_data = default.val("odor"), zero = default.val("zero"),
  tag = "Name", limits, base_size = 12)
}
\arguments{
\item{odorant}{character vecto; one or several InChIKeys}

\item{response_matrix}{DoOR response matrix; a DoOR response matrix as data source}

\item{odor_data}{data frame; contains the odorant information.}

\item{zero}{character; an InChIKey of the odorant that should be set to 0}

\item{tag}{character; the chemical identifier to plot as odorant name (one of colnames(odor))}

\item{limits}{numeric of length 2; if provided the ylim will range accordingly}

\item{base_size}{numeric; the base font size for the ggplot2 plot}
}
\value{
a ggplot object
}
\description{
barplot of DoOR responses of a set of odorant across all responding units in DoOR
}
\examples{
library(DoOR.data)
data(odor)
DoORplot_acrossReceptors(transID("123-92-2"), tag = "CAS")
DoORplot_acrossReceptors(odor$InChIKey[4:10])
}
\author{
Daniel Münch <\email{daniel.muench@uni-konstanz.de}>
}

