% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoORplot_responseMatrix.R
\name{DoORplot_responseMatrix}
\alias{DoORplot_responseMatrix}
\title{DoORplot_responseMatrix}
\usage{
DoORplot_responseMatrix(data, tag = default.val("tag"),
  colors = c("#0570b0", "#74a9cf", "#ffffff", "#fdcc8a", "#fc8d59",
  "#d7301f"), limits)
}
\arguments{
\item{data}{a subset of e.g. response.matrix}

\item{tag}{the chemical identfier to plot (one of colnames(odor))}

\item{colors}{the colors to use if negative values are supplied (range of 5 colors, 2 for negative values, 1 for 0 and 3 for positive values)}

\item{limits}{the limits of the scale, will be calculated if not set}
}
\value{
a dotplot if limits[1] >= 0  or a heatmap if limits[1] < 0
}
\description{
plot DoOR responses as a point matrix
}
\examples{
tmp <- apply(response.matrix, 2, function(x) resetSFR(x,x["SFR"]))
DoORplot_responseMatrix(tmp[10:50,], tag = "Name", limits = range(tmp, na.rm = T))
DoORplot_responseMatrix(response.matrix[10:50,], tag = "Name", limits = range(response.matrix, na.rm = T))
}

