% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoORplot_responseProfile.R
\name{DoORplot_responseProfile}
\alias{DoORplot_responseProfile}
\title{DoORplot_responseProfile}
\usage{
DoORplot_responseProfile(receptor,
  response_matrix = default.val("response.matrix"), tag = default.val("tag"),
  colored = TRUE, colors = default.val("colors"), limits,
  zero = default.val("zero"), scalebar = default.val("scalebar"),
  base_size = 12)
}
\arguments{
\item{receptor}{character; receptor name, any of colnames(response.matrix)}

\item{response_matrix}{a DoOR response.matrix}

\item{tag}{character; chemical identifier for annotation}

\item{colored}{logical; color code the bars according to the response value?}

\item{colors}{character vector; a vector of 5 colors (2 for values < 0, 1 value for 0 and 3 values > 0)}

\item{limits}{numeric of length 2; the limits for the colorscale and the x axis, global range of data will be used if empty}

\item{zero}{character; the odorant response that is set to 0, defaults to "SFR"}

\item{scalebar}{logical; add or suppress scalebars}

\item{base_size}{numeric; the base font size for the ggplot2 plot}
}
\value{
ggplot2 plot
}
\description{
create a barplot of a DoOR response profile
}
\examples{
DoORplot_responseProfile(response.matrix, "Or22a")
DoORplot_responseProfile(response.matrix, "Or22a", tag = "Name")
}
\author{
Daniel Münch <\email{daniel.muench@uni-konstanz.de}>
}

