% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoORplot_tuningCurve.R
\name{DoORplot_tuningCurve}
\alias{DoORplot_tuningCurve}
\title{DoORplot_tuningCurve}
\usage{
DoORplot_tuningCurve(receptor, odorant, response.vector,
  response_matrix = default.val("response.matrix"),
  zero = default.val("zero"), fill.receptor = default.val("color.receptor"),
  fill.odorant = default.val("color.odorant"), odor.main = "Name", limits,
  base_size = 12)
}
\arguments{
\item{receptor}{character; a receptor name (one of ORS$OR)}

\item{odorant}{character; an odorant name (InChIKey)}

\item{response.vector}{numerical vector; a vector with responses, if empty this is taken from response_matrix}

\item{response_matrix}{DoOR response matrix; response vector will be taken from here, not needed if response.vector is given}

\item{zero}{InChIKey; will be set to zero, default is SFR, ignored when data is provided via response.vector; set to "" if you don't want to subtract anything}

\item{fill.receptor}{color code; bar color for receptor tuning curve}

\item{fill.odorant}{color code; bar color for odorant tuning curve}

\item{odor.main}{the odor identifier to plot, one of colnamed(odor)}

\item{limits}{the numerical vector of length 2; y limits for the tuning curve}

\item{base_size}{numeric; the base font size for the ggplot2 plot}
}
\value{
a ggplot object
}
\description{
plot a receptor or odorant tuning curve
}
\examples{
DoORplot_tuningCurve(odorant = odor$InChIKey[2])
DoORplot_tuningCurve(receptor = "Or22a")

range <- range(apply(response_matrix, 2, function(x) resetSFR(x,x[zero])), na.rm = TRUE)
DoORplot_tuningCurve(receptor = "Or10a", limits = range, fill.receptor = "magenta")

DoORplot_tuningCurve(receptor = "OrX", response.vector = c(1:100))

DoORplot_tuningCurve(odorant = "odor X", response.vector = rnorm(200))
}
\author{
Daniel Münch <\email{daniel.muench@uni-konstanz.de}>
}

