% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot_compare_profiles.R
\name{dplot_compare_profiles}
\alias{dplot_compareProfiles}
\alias{dplot_compare_profiles}
\title{Compare two response profiles}
\usage{
dplot_compare_profiles(x, y, by.x, by.y, tag = "Name", base_size = 12)
}
\arguments{
\item{x}{the input data frame the first response profile will be taken from}

\item{y}{the input data frame the second response profile will be taken from
(x will be taken if y is missing)}

\item{by.x}{character string, specifying a column in x}

\item{by.y}{character string, specifying a column in y}

\item{tag}{character, the chemical identifier that will be used as odorant
label.}

\item{base_size}{numeric, the base font size for the ggplot2 plot}
}
\description{
Orderdered bar plots for two studies, allowing for an easy comparison of the
two studies / response profiles'.
}
\examples{
library(DoOR.data)
data(Or22a)
data(response.range)
data(response.matrix)

dplot_compare_profiles(x = Or22a, y = Or22a,
                         by.x = "Hallem.2006.EN",
                         by.y = "Pelz.2006.AntEC50")

dplot_compare_profiles(x = cbind(response.matrix, InChIKey = rownames(response.matrix)),
                         y = cbind(response.matrix, InChIKey = rownames(response.matrix)),
                         by.x = "Or22a", by.y = "Or10a")

}
\author{
Daniel Münch <\email{daniel.muench@uni-konstanz.de}>
}

