% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_door_data.R
\name{export_door_data}
\alias{exportData}
\alias{export_door_data}
\title{export data}
\usage{
export_door_data(file.format, directory,
  odorantReceptors = door_default_values("ORs"),
  response_matrix = door_default_values("door_response_matrix"),
  responseRange = door_default_values("door_response_range"),
  unglobalNorm_RM = door_default_values("door_response_matrix_non_normalized"),
  weightGlobNorm = door_default_values("door_global_normalization_weights"), all.data = TRUE)
}
\arguments{
\item{file.format}{character string, the format of given file, either ".txt"
or ".csv"}

\item{directory}{character string, naming a directory for writing. If
missing, the exported data are saved in current working directory.}

\item{odorantReceptors}{data frame, receptor names and expressions}

\item{response_matrix}{data matrix, an global unnormalized responses matrix}

\item{responseRange}{data frame, response ranges for each study}

\item{unglobalNorm_RM}{data matrix, an unnormalized responses matrix}

\item{weightGlobNorm}{data frame, weight matrix for global normalizazion}

\item{all.data}{logical, if TRUE, export odorant response data and supported
data "door_response_matrix", "door_response_range", "door_response_matrix_non_normalized",
"door_response_matrix", "door_global_normalization_weights" and "ORs".}
}
\description{
export odor response data and supported data
}
\details{
Please load ORs from data package DoOR.data by typing (\code{data(ORs)})
before use.
}
\examples{
\dontrun{
library(DoOR.data)
library(DoOR.functions)
load_door_data()
export_door_data(".txt", all.data = FALSE) 	# export odorant responses data only
}
}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>
}
\keyword{data}

