% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filterData.R
\name{filterData}
\alias{filterData}
\title{filterData}
\usage{
filterData(data, overlapValues = default.val("overlapValues"),
  charColumns = default.val("num.charColumns"))
}
\arguments{
\item{data}{a DoOR data frame (e.g. Or22a)}

\item{overlapValues}{number of minimum overlap needed between data sets}

\item{charColumns}{number of the character columns}
}
\value{
a list containing the filtered data set and a data frame naming the removed data sets and the reason for removal
}
\description{
Filter DoOR datasets before merging (remove studies with few overlapping points or SD = 0). This function is used by updateDatabase to pre filter before performing all possible merges.
}
\examples{
tmp <- filterData(Or42b)
tmp$excluded
}

