% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_normalized_responses.R
\name{get_normalized_responses}
\alias{get_normalized_responses}
\title{Find normalised receptor responses}
\usage{
get_normalized_responses(odors, zero = door_default_values("zero"),
  response_matrix = door_default_values("response.matrix"), round = 3,
  na.rm = FALSE)
}
\arguments{
\item{odors}{character vector, one or more odors provided as InChIKey.}

\item{zero}{InChIKey of background that should be set to zero. The default is
"SFR", i.e. the spontaneous firing rate.}

\item{response_matrix}{a data frame, as e.g. "response.matrix" that is loaded
by \code{\link{modelRP}}. It is also possible to create this frame manually
using \code{\link{modelRP}}.}

\item{round}{numeric, round to this amount of digits, set to NA if you do not want to round}

\item{na.rm}{logical, remove NAs?}
}
\description{
given a chemical, get normalised receptor responses from all studies in the
database.
}
\examples{

library(DoOR.data)
odors <- c("MLFHJEHSLIIPHL-UHFFFAOYSA-N",
           "OBNCKNCVKJNDBV-UHFFFAOYSA-N",
           "IKHGUXGNUITLKF-UHFFFAOYSA-N")
data(response.matrix)
result <- get_normalized_responses(odors, response_matrix = response.matrix)

}
\seealso{
\code{\link{modelRP}},\code{\link{create_door_database}}
}
\keyword{data}

