% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load2list.R
\name{load2list}
\alias{load2list}
\title{load2list}
\usage{
load2list()
}
\value{
a list
}
\description{
returns all original DoOR response data as a list
}
\examples{
# load DoOR.data
library(DoOR.data)
load_door_data(nointeraction = TRUE)

# write the data into a list
lst <- load2list()
}
\author{
Daniel Münch <\email{daniel.muench@uni-konstanz.de}>
}

