% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load2list.R
\name{load2list}
\alias{load2list}
\title{load2list}
\usage{
load2list()
}
\value{
a list
}
\description{
returns all original DoOR response data as a list
}
\examples{
library(DoOR.data)
loadData()
lst <- load2list()
}
\author{
Daniel Münch <\email{daniel.muench@uni-konstanz.de}>
}

