% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadRDList.R
\name{loadRDList}
\alias{loadRDList}
\title{load data into a list}
\usage{
loadRDList(odorantReceptors = default.val("ORs"))
}
\arguments{
\item{odorantReceptors}{data frame; containing receptor or ORN names and
their expression (2: both in adult and larvae; 1: only in larvae, 0: only
in adult). Output is a list containing all measurements from all available
studies.}
}
\description{
load all available data into a list
}
\details{
Please load ORs from data package DoOR.data by typing (\code{data(ORs)})
before use.
}
\examples{
library(DoOR.data)
library(DoOR.function)
loadRD()
orlist <- loadRDList()
orlist$Or22a
}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>
}
\keyword{data}

