% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapReceptor.R
\name{mapReceptor}
\alias{mapReceptor}
\title{mapReceptor}
\usage{
mapReceptor(data, response_matrix = default.val("response.matrix"), sub,
  threshold.p, threshold.cor, nshow)
}
\arguments{
\item{data}{data frame, containing two columns, one called "odorants" and one
"responses" providing InChIKeys and odorant responses respectively.}

\item{response_matrix}{output is a numeric vector that contains the Pearson
Correlation Coefficient between given data and selected consensus data in}

\item{sub}{character, a subset of responding units
returned response matrix}

\item{threshold.p}{numeric, a p-value threshold, only correlations below will be returned}

\item{threshold.cor}{numeric, a correlation-coefficient threshold, only correlations above will be returned}

\item{nshow}{numeric, if defined, only this number of results will be}
}
\description{
Identifying the source of unknown response data by correlating it agains all
DoOR responding units.
}
\examples{
loadData()
  data <- data.frame(odorants  = Or22a$InChIKey,
                     responses = Or22a$Hallem.2004.EN)
  data <- na.omit(data)
  mapReceptor(data = data)
}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>

Daniel Münch <\email{daniel.muench@uni-konstanz.de}>
}

