% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modelfunction_asymp.R
\name{modelfunction_asymp}
\alias{modelfunction_asymp}
\title{Asymptotic Model}
\usage{
modelfunction_asymp(input, parms)
}
\arguments{
\item{input}{numeric vector of data values}

\item{parms}{numeric vector; parameters with given names: "Asym", "R0",
"lrc"}
}
\description{
Asymptotic Model
}
\details{
expression : Asym+(R0-Asym)*exp(-exp(lrc)*input)
}
\examples{
x <- rnorm(20)
y <- modelfunction_asymp(input = x, parms = c(Asym = 0.1, R0 = 0.2, lrc = 0.3) )
plot(x,y)
}
\references{
Jose Pinheiro and Douglas Bates
}
\seealso{
\code{\link{SSasymp}}
}
\keyword{math}

