% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelfunction_sigmoid.R
\name{modelfunction_sigmoid}
\alias{modelfunction_sigmoid}
\title{Logistic Model}
\usage{
modelfunction_sigmoid(input, parms)
}
\arguments{
\item{input}{numeric vector of data values}

\item{parms}{numeric vector; parameters with given names: "Asym", "xmid",
"scal"}
}
\description{
Logistic Model
}
\details{
expression : Asym/(1+exp((xmid-input)/scal))
}
\examples{
x <- rnorm(20)
y <- modelfunction_sigmoid(input = x, parms = c(Asym = 0.1, xmid = 0.2, scal = 0.3) )
plot(x,y)
}
\references{
Jose Pinheiro and Douglas Bates
}
\seealso{
\code{\link{SSlogis}}
}
\keyword{math}

