% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelRPSEQ.R
\name{modelRPSEQ}
\alias{modelRPSEQ}
\title{modelRPSEQ}
\usage{
modelRPSEQ(data, SEQ, overlapValues = door_default_values("overlapValues"),
  select.MDValue = door_default_values("select.MDValue"), strict = TRUE,
  plot = FALSE)
}
\arguments{
\item{data}{data frame, odorant response data, e.g. Or22a.}

\item{SEQ}{character vector, containing the names of studies indicating given
sequence for merging data.}

\item{overlapValues}{minimum overlap between studies to perfom a merge}

\item{select.MDValue}{the minimum mean distance between studies to perfom a
merge}

\item{strict}{logical, if TRUE merging a permutation will be stopped once a
single merge has a mean distance above select.MDValue}

\item{plot}{logical}
}
\description{
generates a model response and merge data in given sequence
}
\details{
# modelRPSEQ.R: #################

# merges studies in a given sequence (determined by the user or by exhaustive
enumeration and choosing the optimal sequence)


# input parameters: ####################


# data  : data frame, odorant response data for a given receptor, e.g. Or22a
# SEQ 	: character vector, contains the names of studies that measured this
receptor in a specific order (the merging sequence)

# output is a numeric vector: response values
}
\examples{

library(DoOR.data)
data(Or35a)
data(response.range)
SEQ <- c("Hallem.2006.EN","Kreher.2008.EN","Hallem.2006.EN")
selected.merg <- modelRPSEQ(Or35a, SEQ = SEQ, plot = TRUE)

}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>
}
\keyword{data}

