% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PseudoInverse.R
\name{PseudoInverse}
\alias{PseudoInverse}
\title{Evaluate the pseudoinvese of a matrix}
\usage{
PseudoInverse(x)
}
\arguments{
\item{x}{a matrix}
}
\description{
Evaluate the pseudoinvese of a matrix
}
\details{
The pseudoinverse of the matrix x with singular value decomposition \eqn{x=U
D V^'} is: \eqn{x^+ = V D^+ U^'} where D^+ is the pseudoinverse of D, which
is formed by replacing every nonzero entry by its reciprocal.
}
\examples{
x  <- matrix(seq(1,2,length=4),nrow = 2, ncol = 2)
PI <- PseudoInverse(x)
x
x\%*\%PI\%*\%x
}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>
}
\seealso{
\code{\link{LLSIestPC}}, \code{\link{svd}}
}
\keyword{math}

