% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_sfr.R
\name{reset_sfr}
\alias{reset_sfr}
\alias{resetSFR}
\title{reset SFR}
\usage{
reset_sfr(x, sfr)
}
\arguments{
\item{x}{numeric or DoOR response matrix, input values}

\item{sfr}{numeric or character, either a value to subtract if x is a vector or an InChIKey if x is a DoOR response matrix}
}
\description{
A function for reseting SFR to zero
}
\details{
Performs a simple subtraction of the SFR value.
}
\examples{
# load data
library(DoOR.data)
data(door_response_matrix)

# create a response matrix with the SFR reset to 0
door_response_matrix_SFRreset <- reset_sfr(door_response_matrix, "SFR")

}
\author{
Daniel Münch <\email{daniel.muench@uni-konstanz.de}>
}
