% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/resetSFR.R
\name{resetSFR}
\alias{resetSFR}
\title{reset SFR}
\usage{
resetSFR(x, sfr)
}
\arguments{
\item{x}{a numeric vector of input values.}

\item{sfr}{a numeric parameter, indicating spontaneous firing rate value.}
}
\description{
A function for reseting SFR to zero
}
\details{
Performs a simple subtraction of the SFR value.
}
\examples{
library(DoOR.data)
data(response.matrix)
response.matrix.SFRreset <- apply(response.matrix, 2, function(x) resetSFR(x,x['SFR']))
}
\author{
Daniel Münch <\email{daniel.muench@uni-konstanz.de}>
}
\keyword{math}

