% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/resetSFR.R
\name{resetSFR}
\alias{resetSFR}
\title{reset SFR}
\usage{
resetSFR(x, sfr)
}
\arguments{
\item{x}{numeric or DoOR response matrix, input values}

\item{sfr}{numeric or character, either a value to subtract if x is a vector or an InChIKey if x is a DoOR response matrix}
}
\description{
A function for reseting SFR to zero
}
\details{
Performs a simple subtraction of the SFR value.
}
\examples{
library(DoOR.data)
data(response.matrix)
response.matrix.SFRreset <- resetSFR(response.matrix, "SFR")
}
\author{
Daniel Münch <\email{daniel.muench@uni-konstanz.de}>
}

