% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/showOdor.R
\name{showOdor}
\alias{showOdor}
\title{Link to PubChem}
\usage{
showOdor(CID, CAS, odor.data = default.val("odor.data"))
}
\arguments{
\item{CID}{a character string; the compound identity (CID) in pubchem.}

\item{CAS}{a character string; the CAS number.}

\item{odor.data}{data frame; chemical data of odors.}
}
\description{
obtain information on a chemical from PubChem
}
\details{
Send a query to PubChem. The result is displayed using the default
WWW-browser.
}
\examples{
library(DoOR.data)
data(odor)
showOdor(CID=222,odor.data=odor)
showOdor(CAS="64-19-7",odor.data=odor)
}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>
}
\seealso{
\code{\link{browseURL}}
}
\keyword{data}

