% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SSexpo.R
\docType{data}
\name{SSexpo}
\alias{SSexpo}
\title{SelfStart estimation of exponential function}
\format{\preformatted{function (x, a, b, c)  
 - attr(*, "srcref")=Class 'srcref'  atomic [1:8] 21 11 36 1 11 1 21 36
  .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x45b9038> 
 - attr(*, "initial")=function (mCall, data, LHS)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 36 14 65 3 14 3 36 65
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x45b9038> 
 - attr(*, "pnames")= chr [1:3] "a" "b" "c"
 - attr(*, "class")= chr "selfStart"
}}
\usage{
SSexpo
}
\arguments{
\item{x}{a numeric vector indicates input values in an equation.}

\item{a}{a numieric parameter.}

\item{b}{a numieric parameter.}

\item{c}{a numieric parameter.}
}
\description{
selfStart model estimates the parameters (a,b,c) of expontenial model.
}
\details{
The expression of eponential function is \code{a+b*exp(c*x)}.
}
\examples{
x <- -(1:100)/10
y <- 1 + 10 * exp(x / 2) + rnorm(x)/10
pr<-nls(y~SSexpo(x,a,b,c))
}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>
}
\seealso{
\code{\link{nls}}, \code{\link{selfStart}}
}
\keyword{math}

