% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_odorinfo.R
\name{update_odorinfo}
<<<<<<< HEAD:man/update_odorinfo.Rd
\alias{updateOdorInfo}
=======
>>>>>>> 9d73a9bc3fe04986e1b96d7d4c8b786026637987:man/updateOdorInfo.Rd
\alias{update_odorinfo}
\title{update_odorinfo}
\usage{
update_odorinfo()
}
\description{
Update the DoOR odor data with info from \code{odor}. For the function to
work, all DoOR data has to be loaded.
}
\author{
Daniel Münch, \email{daniel@muench.bio}
}

