% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/updateDatabase.R
\name{updateDatabase}
\alias{updateDatabase}
\title{update response matrix}
\usage{
updateDatabase(receptor, permutation = TRUE, perm,
  responseMatrix_nn = default.val("response.matrix_non.normalized"),
  responseMatrix = default.val("response.matrix"),
  responseRange = default.val("response.range"),
  weightGlobNorm = default.val("weight.globNorm"),
  select.MDValue = default.val("select.MDValue"), strict = TRUE,
  overlapValues = default.val("overlapValues"),
  excluded.data = default.val("excluded.data"), plot = FALSE)
}
\arguments{
\item{receptor}{character string; name of given odorant receptor.}

\item{permutation}{logical; if TRUE, the sequence is chosen from
permutation, if FALSE, sequence is chosen by the routine process.}

\item{perm}{a matrix with one sequence of study names per row, if empty, all possible permutations of study names will be provided.}

\item{responseMatrix_nn}{data frame; response data that has not
been globally normalized.}

\item{responseMatrix}{data frame; globally normalized response data.}

\item{responseRange}{data frame; response range of studies.}

\item{weightGlobNorm}{data frame; weight matrix for global normalization.}

\item{select.MDValue}{the minimum mean distance between studies to perfom a merge (used if permutation == FALSE or if permutation == TRUE AND strict == TRUE)}

\item{strict}{logical, if TRUE merging a permutation will be stopped once a single merge has a mean distance above select.MDValue (only valid if permutation == TRUE)}

\item{overlapValues}{minimum overlap between studies to perfom a merge}
}
\description{
update the globally \code{response matrix} and the unglobally normalized
response matrix \code{response.matrix_non.normalized} by introducing new
consensus response data of given receptor.
}
\details{
The merging sequence could be arranged by the routine process (using
\code{\link{modelRP}} or taking the optimized sequence that is chosen from
permutations. The mean correlation between merged responses and each
original recording will be computed for each permutation, the optimozed
sequence is with the highest correlation.
}
\examples{
\dontrun{
library(DoOR.data)
loadData()
# update the entry "Or67b" of data "response.matrix" and "response.matrix_non.normalized" with permutations.
# updateDatabase(receptor="Or67b", permutation = TRUE)
}
}
\author{
Shouwen Ma <\email{shouwen.ma@uni-konstanz.de}>

Shouwen Ma <\email{daniel.muench@uni-konstanz.de}>
}
\seealso{
\code{\link{modelRP}},\code{\link{modelRPSEQ}}
}
\keyword{data}

