% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateOdorInfo.R
\name{updateOdorInfo}
\alias{updateOdorInfo}
\title{updateOdorInfo}
\usage{
updateOdorInfo()
}
\description{
Update the DoOR odor data with info from \code{odor}. For the function to
work, all DoOR data has to be loaded.
}
\author{
Daniel Münch, \email{daniel@muench.bio}
}

